Add item handlers and service

+++ repositories/ItemRepository.go
@@ -12,10 +12,10 @@ type itemRepository struct {
 
 type ItemRepository interface {
 	GetAllItems(ctx context.Context) ([]*entities.Item, error)
-	GetItem(ctx context.Context, id int64) (*entities.Item, error)
+	GetItem(ctx context.Context, id int) (*entities.Item, error)
 	CreateItem(ctx context.Context, item *entities.Item) error
 	UpdateItem(ctx context.Context, item *entities.Item) error
-	DeleteItem(ctx context.Context, id int64) error
+	DeleteItem(ctx context.Context, id int) error
 }
 
 func InitItemRepository(db *sqlx.DB) ItemRepository {
@@ -28,7 +28,7 @@ func (i itemRepository) GetAllItems(ctx context.Context) ([]*entities.Item, erro
 	return items, err
 }
 
-func (i itemRepository) GetItem(ctx context.Context, id int64) (*entities.Item, error) {
+func (i itemRepository) GetItem(ctx context.Context, id int) (*entities.Item, error) {
 	var item entities.Item
 	err := i.db.GetContext(ctx, &item, "CALL sp_GetItem(?)", id)
 	return &item, err
@@ -44,7 +44,7 @@ func (i itemRepository) UpdateItem(ctx context.Context, item *entities.Item) err
 	return err
 }
 
-func (i itemRepository) DeleteItem(ctx context.Context, id int64) error {
+func (i itemRepository) DeleteItem(ctx context.Context, id int) error {
 	_, err := i.db.ExecContext(ctx, "CALL sp_DeleteItem(?)", id)
 	return err
 }
